consolidate utils
